added yes cancel validation on delete (crud.php)

+++ crud.php
@@ -81,7 +81,7 @@
                     echo "</td>";
 
                     echo "<td>";
-                    echo "<a href='delete.php?id=".$result[$i]['id_olahraga']."'>Delete</a>";
+                    echo "<a href='delete.php?id=".$result[$i]['id_olahraga']."' onclick=\"return confirm('apakah anda yakin?')\" >Delete</a>";
                     echo "</td>";
                     echo "</tr>";
                 }
